requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
python-dotenv
plaid-python
django-celery-beat
# Security packages
bandit
safety
# Heroku deployment packages
gunicorn
psycopg2-binary
whitenoise
redis
dj-database-url